v1.0.0

+++ PharmacyManagermentSystem/Models/Salary.cs
@@ -9,8 +9,8 @@ namespace PharmacyManagermentSystem.Model
     {
         public int Month { get; set; }
         public int Year { get; set; }
-        public int BasicSalary { get; set; }
-        public int Bonus { get; set; }
+        public double BasicSalary { get; set; }
+        public double Bonus { get; set; }
         public int DayWorked { get; set; }
         public int DayOff { get; set; }
         public string EmployeeId { get; set; }
