+++ PharmacyManagermentSystem/Request/ReceiptDetailRequest.cs
@@ -11,7 +11,7 @@ namespace PharmacyManagermentSystem.Request
         public string CategoryId { get; set; }
         public int Quantity { get; set; }
         public string Status { get; set; }
-        public int Price { get; set; }
+        public double Price { get; set; }
     }
     public class UpdateReceiptDetailRequest
     {
@@ -22,7 +22,7 @@ namespace PharmacyManagermentSystem.Request
         public string NewCategoryId { get; set; }
         public int Quantity { get; set; }
         public string Status { get; set; }
-        public int Price { get; set; }
+        public double Price { get; set; }
     }
     public class DeleteReceiptDetailRequest
     {
