+++ PharmacyManagermentSystem/Request/SalaryRequest.cs
@@ -4,8 +4,8 @@
     {
         public int Month { get; set; }
         public int Year { get; set; }
-        public int BasicSalary { get; set; }
-        public int Bonus { get; set; }
+        public double BasicSalary { get; set; }
+        public double Bonus { get; set; }
         public int DayWorked { get; set; }
         public int DayOff { get; set; }
         public string EmployeeId { get; set; }
@@ -14,8 +14,8 @@
     {
         public int Month { get; set; }
         public int Year { get; set; }
-        public int BasicSalary { get; set; }
-        public int Bonus { get; set; }
+        public double BasicSalary { get; set; }
+        public double Bonus { get; set; }
         public int DayWorked { get; set; }
         public int DayOff { get; set; }
         public string EmployeeId { get; set; }
